+++ MoonshotService.qml
@@ -1,0 +1,9 @@
+import QtQuick
+
+MoonshotModel {
+  id: root
+
+  property string omarchyPath: ""
+  property var shell: null
+  property var manifest: null
+}